eview fixes
