+++ package.json
@@ -15,6 +15,6 @@
     "eslint-config-airbnb-base": "15.0.0",
     "eslint-plugin-import": "2.25.4",
     "js-yaml": "4.0.0",
-    "jwt-decode": "^3.1.2"
+    "jwt-decode": "3.1.2"
   }
 }
